+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [2829](https://github.com/influxdb/influxdb/pull/2829) -- Re-enable Graphite support as a new Service-style component.
 - [2814](https://github.com/influxdb/influxdb/issues/2814) -- Convert collectd to a service.
 - [2852](https://github.com/influxdb/influxdb/pull/2852) -- Don't panic when altering retention policies. Thanks for the report @huhongbo
+- [2857](https://github.com/influxdb/influxdb/issues/2857) -- Fix parsing commas in string field values.
 
 ## v0.9.0-rc32 [2015-06-07]